using size_t

+++ util/dynamic_bloom.h
@@ -123,7 +123,7 @@ inline bool DynamicBloom::MayContain(const Slice& key) const {
 inline void DynamicBloom::MayContain(int num_keys, Slice** keys,
                                      bool* may_match) const {
   std::array<uint32_t, MultiGetContext::MAX_BATCH_SIZE> hashes;
-  std::array<uint32_t, MultiGetContext::MAX_BATCH_SIZE> byte_offsets;
+  std::array<size_t, MultiGetContext::MAX_BATCH_SIZE> byte_offsets;
   for (int i = 0; i < num_keys; ++i) {
     hashes[i] = BloomHash(*keys[i]);
     size_t a = fastrange32(kLen, hashes[i]);
